Remove unsemantic, unneeded `div#content`
This is something that I don't need yet, so I'm going to take it out.
It was not being styled.  And, most importantly, it interferes with a
main tag that functions as a `content-grid` (since that CSS assumes
that everything inside the content grid is a direct child of the
element with the content grid applied).