Define font-related variables
This is needed if we want to use the new layout/CSS stuff with
different fonts anywhere.  Basically we never finished building out
these font styling capabilities since we were only building towards
one page, and that page didn't need anything fancy in terms of fonts.
But now: we do.